requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.10

# IFC Processing
ifcopenshell>=0.7.0
pandas>=2.0.0

# HTTP Client for events
httpx>=0.27.0

# PDF Generation
reportlab>=4.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support
asyncio>=3.4.3
